AP: Bugfix: Altitude leveling value has going lost, if the altitude  unit was changed during leveling. Furthermore was the correction value for unit feet never saved.
git-svn-id: https://svn.kflog.org/svn/repos/cumulus/qt4/trunk@6589 c8f415cb-fa42-0410-b21e-89df95f64ae5

+++ cumulus/generalconfig.cpp
@@ -490,7 +490,7 @@ void GeneralConfig::load()
   _gpsBtDevice        = value( "BT-Device", "" ).toString();
   _gpsSpeed           = value( "Speed", 4800 ).toInt();
   _gpsAltitudeType    = value( "AltitudeType", (int) GpsNmea::GPS ).toInt();
-  _gpsAltitudeUserCorrection.setMeters(value( "AltitudeCorrection", 0 ).toInt());
+  _gpsAltitudeUserCorrection.setMeters(value( "AltitudeCorrection", 0 ).toDouble());
   _gpsSoftStart       = value( "SoftStart", false ).toBool();
   _gpsHardStart       = value( "HardStart", false ).toBool();
   _gpsSyncSystemClock = value( "SyncSystemClock", false ).toBool();
